+++ sqlalchemy_kinetica/custom_commands.py
@@ -1,4 +1,5 @@
-from os import readv
+"""Kinetica custom SQL commands and statement classes."""
+
 from random import randint
 
 from sqlalchemy import Executable, ClauseElement, FunctionElement, literal_column, BinaryExpression, Select
@@ -6,9 +7,15 @@ from sqlalchemy.exc import CompileError
 from sqlalchemy.ext.compiler import compiles
 from sqlalchemy.sql.dml import Insert, Update
 from sqlalchemy.sql.functions import GenericFunction
-from sqlparse.sql import Where
 
 from sqlalchemy_kinetica.dialect import KI_INSERT_HINT_KEY
+
+# Keys for storing custom attributes in kwargs (survives SQLAlchemy cloning)
+# Keys must follow {dialect}_{option} format for SQLAlchemy's _DialectArgDict
+KI_UPDATE_FROM_TABLE_KEY = 'kinetica_from_table'
+KI_UPDATE_JOIN_CONDITION_KEY = 'kinetica_join_condition'
+KI_UPDATE_WHERE_CONDITION_KEY = 'kinetica_where_condition'
+
 
 def quote_qualified_table_name(qualified_table_name):
     if not qualified_table_name:
@@ -17,11 +24,13 @@ def quote_qualified_table_name(qualified_table_name):
     quoted_parts = [f"\"{part}\"" for part in parts]
     return '.'.join(quoted_parts)
 
+
 def safe_index(lst, value):
     return lst.index(value) if value in lst else len(lst)
 
 
 class KiInsert(Insert):
+    """Custom INSERT statement with Kinetica hint support."""
     inherit_cache = False
 
     def __init__(self, *args, insert_hint=None, **kwargs):
@@ -41,23 +50,30 @@ def ki_insert(table, insert_hint=None, **kwargs):
 
 
 class KiUpdate(Update):
-    def __init__(self, table, from_table, join_condition=None, where_condition=None,  *args, **kwargs):
+    """Custom UPDATE statement with Kinetica JOIN/FROM support.
+
+    Stores custom attributes in kwargs to survive SQLAlchemy's cloning mechanism.
+    """
+    inherit_cache = False
+
+    def __init__(self, table, from_table=None, join_condition=None, where_condition=None, *args, **kwargs):
         super().__init__(table, *args, **kwargs)
-        self._from_table = from_table
-        self._join_condition = join_condition
-        self._where_condition = where_condition
+        # Store in kwargs to survive cloning (follows KiInsert pattern)
+        self.kwargs[KI_UPDATE_FROM_TABLE_KEY] = from_table
+        self.kwargs[KI_UPDATE_JOIN_CONDITION_KEY] = join_condition
+        self.kwargs[KI_UPDATE_WHERE_CONDITION_KEY] = where_condition
 
     @property
     def from_table(self):
-        return self._from_table
+        return self.kwargs.get(KI_UPDATE_FROM_TABLE_KEY)
 
     @property
     def join_condition(self):
-        return self._join_condition
+        return self.kwargs.get(KI_UPDATE_JOIN_CONDITION_KEY)
 
     @property
     def where_condition(self):
-        return self._where_condition
+        return self.kwargs.get(KI_UPDATE_WHERE_CONDITION_KEY)
 
 
 @compiles(KiUpdate, 'kinetica')
@@ -66,6 +82,7 @@ def compile_ki_update(update, compiler, **kwargs):
 
 
 class InsertFromSelect(Executable, ClauseElement):
+    """INSERT INTO ... SELECT statement."""
     inherit_cache = False
 
     def __init__(self, table, select):
@@ -82,16 +99,17 @@ def visit_insert_from_select(element, compiler, **kw):
 
 
 class CreateTableAs(Executable, ClauseElement):
+    """CREATE TABLE ... AS SELECT statement."""
     inherit_cache = False
 
-    def __init__(self, table, query, prefixes=None, table_properties= None):
+    def __init__(self, table, query, prefixes=None, table_properties=None):
         self.table = table
         self.query = query
         self.prefixes = prefixes
         self.table_properties = table_properties
 
 
-@compiles(CreateTableAs, "kinetica")
+@compiles(CreateTableAs, 'kinetica')
 def _create_table_as(element, compiler, **kw):
 
     def build_table_properties(d):
@@ -117,11 +135,13 @@ def _create_table_as(element, compiler, **kw):
 
 # Define the custom function
 class Asof(FunctionElement):
+    """ASOF join function for time-series data."""
     inherit_cache = True
     name = 'asof'
 
+
 # Compile the function into SQL
-@compiles(Asof)
+@compiles(Asof, 'kinetica')
 def compile_asof(element, compiler, **kwargs):
     left_column = compiler.process(element.clauses.clauses[0])
     right_column = compiler.process(element.clauses.clauses[1])
@@ -133,6 +153,7 @@ def compile_asof(element, compiler, **kwargs):
 
 
 class FirstValue(GenericFunction):
+    """FIRST_VALUE window function with IGNORE/RESPECT NULLS support."""
     inherit_cache = True
     name = 'first_value'
 
@@ -141,12 +162,13 @@ class FirstValue(GenericFunction):
         self.ignore_nulls = ignore_nulls
 
 
-@compiles(FirstValue, "kinetica")
+@compiles(FirstValue, 'kinetica')
 def compile_first_value(element, compiler, **kwargs):
     return compiler.visit_first_value(element, **kwargs)
 
 
-class Pivot(ClauseElement):
+class Pivot:
+    """Helper class to store PIVOT clause information."""
     def __init__(self, value_agg_fn, type_column, type_values):
         self.value_agg_fn = value_agg_fn
         self.type_column = type_column
@@ -155,17 +177,42 @@ class Pivot(ClauseElement):
 
 # Extend the Select class to support PIVOT
 class PivotSelect(Select):
+    """SELECT statement with PIVOT support.
+
+    Uses instance attributes with _copy_internals override to survive cloning.
+    """
     inherit_cache = False
+
+    _pivot = None
+    _source_alias = None
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._source_alias = None
         self._pivot = None
+        self._source_alias = None
+
+    def _copy_internals(self, clone=None, **kw):
+        """Override to preserve custom attributes during cloning."""
+        super()._copy_internals(clone=clone, **kw)
+        # Note: _copy_internals is called on the NEW (cloned) object
+        # The clone parameter is a function, not the source object
+        # We need to copy from the original, but this method is called on the clone
+        # So we actually don't need to do anything here - see _generate() instead
+
+    def _generate(self):
+        """Override to preserve custom attributes when generating new instances."""
+        cloned = super()._generate()
+        cloned._pivot = self._pivot
+        cloned._source_alias = self._source_alias
+        return cloned
 
     def select_from(self, pivot_source):
-        self._source_alias = f"p_sq_{randint(0,1000000000)}"
-        self = super().select_from(pivot_source.alias(self._source_alias))
-        return self
+        self._source_alias = f"p_sq_{randint(0, 1000000000)}"
+        result = super().select_from(pivot_source.alias(self._source_alias))
+        # Copy custom attributes to the new instance
+        result._pivot = self._pivot
+        result._source_alias = self._source_alias
+        return result
 
     def pivot(self, value_agg_fn, type_column, type_values):
         self._pivot = Pivot(value_agg_fn, type_column, type_values)
@@ -173,22 +220,25 @@ class PivotSelect(Select):
 
 
 # Custom compilation for Select that includes the PIVOT clause
-@compiles(PivotSelect)
+@compiles(PivotSelect, 'kinetica')
 def compile_pivot_select(element, compiler, **kwargs):
 
     # Compile the inner select first
     query = compiler.visit_select(element, **kwargs)
 
     # If there's a pivot clause, inject it
-    if element._pivot is not None:
-        sq_alias_clause = f" AS {element._source_alias}"
+    pivot = element._pivot
+    source_alias = element._source_alias
+
+    if pivot is not None and source_alias is not None:
+        sq_alias_clause = f" AS {source_alias}"
         sq_alias_begin_index = safe_index(query, sq_alias_clause)
         sq_alias_end_index = sq_alias_begin_index + len(sq_alias_clause) + 1
-        
+
         pivot_sql = (
-            f"PIVOT ({element._pivot.value_agg_fn} "
-            f"FOR {element._pivot.type_column} "
-            f"IN ({', '.join(map(str, element._pivot.type_values))}))"
+            f"PIVOT ({pivot.value_agg_fn} "
+            f"FOR {pivot.type_column} "
+            f"IN ({', '.join(map(str, pivot.type_values))}))"
         )
 
         query = f"{query[:sq_alias_begin_index]}\n{pivot_sql}\n{query[sq_alias_end_index:]}"
@@ -198,6 +248,7 @@ def compile_pivot_select(element, compiler, **kwargs):
 
 # Custom UNPIVOT clause class
 class Unpivot:
+    """Helper class to store UNPIVOT clause information."""
     def __init__(self, unpivoted_value_column, unpivoted_type_column, value_columns):
         self.unpivoted_value_column = unpivoted_value_column
         self.unpivoted_type_column = unpivoted_type_column
@@ -206,17 +257,38 @@ class Unpivot:
 
 # Extend the Select class to include the UNPIVOT clause
 class UnpivotSelect(Select):
+    """SELECT statement with UNPIVOT support.
+
+    Uses instance attributes with _generate override to survive cloning.
+    """
     inherit_cache = False
+
+    _unpivot = None
+    _source_alias = None
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._source_alias = None
         self._unpivot = None
+        self._source_alias = None
+
+    def _copy_internals(self, clone=None, **kw):
+        """Override to preserve custom attributes during cloning."""
+        super()._copy_internals(clone=clone, **kw)
+
+    def _generate(self):
+        """Override to preserve custom attributes when generating new instances."""
+        cloned = super()._generate()
+        cloned._unpivot = self._unpivot
+        cloned._source_alias = self._source_alias
+        return cloned
 
     def select_from(self, unpivot_source):
-        self._source_alias = f"up_sq_{randint(0,1000000000)}"
-        self = super().select_from(unpivot_source.alias(self._source_alias))
-        return self
+        self._source_alias = f"up_sq_{randint(0, 1000000000)}"
+        result = super().select_from(unpivot_source.alias(self._source_alias))
+        # Copy custom attributes to the new instance
+        result._unpivot = self._unpivot
+        result._source_alias = self._source_alias
+        return result
 
     def unpivot(self, unpivoted_value_column, unpivoted_type_column, value_columns):
         self._unpivot = Unpivot(unpivoted_value_column, unpivoted_type_column, value_columns)
@@ -224,22 +296,25 @@ class UnpivotSelect(Select):
 
 
 # Custom SQL compilation for the Unpivot clause
-@compiles(UnpivotSelect)
+@compiles(UnpivotSelect, 'kinetica')
 def compile_unpivot_select(element, compiler, **kwargs):
 
     # Compile the inner select first
     query = compiler.visit_select(element, **kwargs)
 
     # If there's an unpivot clause, inject it
-    if element._unpivot is not None:
-        sq_alias_clause = f" AS {element._source_alias}"
+    unpivot = element._unpivot
+    source_alias = element._source_alias
+
+    if unpivot is not None and source_alias is not None:
+        sq_alias_clause = f" AS {source_alias}"
         sq_alias_begin_index = safe_index(query, sq_alias_clause)
         sq_alias_end_index = sq_alias_begin_index + len(sq_alias_clause) + 1
 
         unpivot_sql = (
-            f"UNPIVOT ({element._unpivot.unpivoted_value_column} "
-            f"FOR {element._unpivot.unpivoted_type_column} "
-            f"IN ({', '.join(element._unpivot.value_columns)}))"
+            f"UNPIVOT ({unpivot.unpivoted_value_column} "
+            f"FOR {unpivot.unpivoted_type_column} "
+            f"IN ({', '.join(unpivot.value_columns)}))"
         )
 
         query = f"{query[:sq_alias_begin_index]}\n{unpivot_sql}\n{query[sq_alias_end_index:]}"
@@ -248,7 +323,10 @@ def compile_unpivot_select(element, compiler, **kwargs):
 
 
 class FilterByString(Executable, ClauseElement):
-    def __init__(self, table_name, mode, expression, view_name = None, column_names = None, options = None):
+    """FILTER_BY_STRING table function."""
+    inherit_cache = False
+
+    def __init__(self, table_name, mode, expression, view_name=None, column_names=None, options=None):
         self.execution_mode = "EXECUTE" if view_name else "SELECT"
         self.table_name = table_name
         self.view_name = view_name
@@ -261,7 +339,7 @@ class FilterByString(Executable, ClauseElement):
 
 
 # Custom SQL compiler for FilterByString
-@compiles(FilterByString)
+@compiles(FilterByString, 'kinetica')
 def compile_filter_by_string(element, compiler, **kw):
     table_name = compiler.process(element.table_name, **kw)
     mode = compiler.process(literal_column(element.mode), **kw)
@@ -297,7 +375,10 @@ def compile_filter_by_string(element, compiler, **kw):
 
 
 class EvaluateModel(Executable, ClauseElement):
-    def __init__(self, model, deployment_mode, replications, source_table, destination_table = None):
+    """EVALUATE_MODEL table function for ML model evaluation."""
+    inherit_cache = False
+
+    def __init__(self, model, deployment_mode, replications, source_table, destination_table=None):
         self.execution_mode = "EXECUTE" if destination_table else "SELECT"
         self.model = model
         self.deployment_mode = deployment_mode
@@ -307,7 +388,7 @@ class EvaluateModel(Executable, ClauseElement):
 
 
 # Custom SQL compiler for EvaluateModel
-@compiles(EvaluateModel)
+@compiles(EvaluateModel, 'kinetica')
 def compile_evaluate_model(element, compiler, **kw):
     # Extract the arguments passed to the function
     model = compiler.process(literal_column(element.model), **kw)